Move pick symbol lengths to utils, clarify comments
